requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
sqlalchemy